tracee-ebpf: extract logic into pkg/cmd
Extracts the tracee-ebpf logic into pkg/cmd. This package allows the use
of the same logic on other places, and also breaks flags parsing and
running tracee. Allowing us to experiment with a new flag library.
