+++ pkg/cmd/flags/capabilities.go
