+++ pkg/cmd/flags/capture.go
